# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "FeedbackAdjudication": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "note": {
            "maxLength": 4000,
            "minLength": 3,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "note"
        ],
        "title": "FeedbackAdjudication",
        "type": "object"
      },
      "FeedbackCreate": {
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "label": {
            "$ref": "#/components/schemas/FeedbackLabel"
          },
          "rationale": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "label",
          "confidence",
          "rationale"
        ],
        "title": "FeedbackCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestigationCreate": {
        "properties": {
          "alert_id": {
            "format": "uuid",
            "title": "Alert Id",
            "type": "string"
          },
          "assigned_to": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "priority": {
            "$ref": "#/components/schemas/InvestigationPriority",
            "default": "MEDIUM"
          },
          "sla_hours": {
            "default": 24,
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Sla Hours",
            "type": "integer"
          },
          "snapshot_id": {
            "format": "uuid",
            "title": "Snapshot Id",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "alert_id",
          "snapshot_id",
          "title"
        ],
        "title": "InvestigationCreate",
        "type": "object"
      },
      "InvestigationEventCreate": {
        "properties": {
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "event_type": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Event Type",
            "type": "string"
          }
        },
        "required": [
          "event_type"
        ],
        "title": "InvestigationEventCreate",
        "type": "object"
      },
      "InvestigationPriority": {
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "URGENT"
        ],
        "title": "InvestigationPriority",
        "type": "string"
      },
      "InvestigationStatus": {
        "enum": [
          "OPEN",
          "TRIAGE",
          "INVESTIGATING",
          "AWAITING_REVIEW",
          "CLOSED"
        ],
        "title": "InvestigationStatus",
        "type": "string"
      },
      "InvestigationUpdate": {
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "disposition": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disposition"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "InvestigationUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Event-time-aware pump-and-dump intelligence backend.",
    "title": "Scam2Market Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/alerts/{alert_id}/evidence": {
      "get": {
        "operationId": "alert_evidence_api_v1_alerts__alert_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Alert Evidence Api V1 Alerts  Alert Id  Evidence Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alert Evidence",
        "tags": [
          "evidence",
          "investigations"
        ]
      }
    },
    "/api/v1/alerts/{alert_id}/explanation": {
      "get": {
        "operationId": "alert_explanation_api_v1_alerts__alert_id__explanation_get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Alert Explanation Api V1 Alerts  Alert Id  Explanation Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alert Explanation",
        "tags": [
          "evidence",
          "investigations"
        ]
      }
    },
...
    "/api/v1/evidence/{snapshot_id}/manifest": {
      "get": {
        "operationId": "evidence_manifest_api_v1_evidence__snapshot_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Access-Reason",
            "required": true,
            "schema": {
              "minLength": 5,
              "title": "X-Access-Reason",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Evidence Manifest Api V1 Evidence  Snapshot Id  Manifest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence Manifest",
        "tags": [
          "evidence",
          "investigations"
        ]
      }
    },
...
    "/api/v1/feedback/{feedback_id}/adjudicate": {
      "post": {
        "operationId": "adjudicate_feedback_api_v1_feedback__feedback_id__adjudicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "feedback_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Feedback Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackAdjudication"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Adjudicate Feedback Api V1 Feedback  Feedback Id  Adjudicate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adjudicate Feedback",
        "tags": [
          "evidence",
          "investigations"
        ]
      }
    },
...
    "/api/v1/investigations": {
      "get": {
        "operationId": "list_investigations_api_v1_investigations_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope_id",
            "required": false,
            "schema": {
              "default": "LIVE",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvestigationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "assigned_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned To"
            }
          },
          {
            "in": "query",
            "name": "overdue_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Overdue Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Investigations Api V1 Investigations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Investigations",
        "tags": [
          "evidence",
          "investigations"
        ]
      },
      "post": {
        "operationId": "create_investigation_api_v1_investigations_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Investigation Api V1 Investigations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Investigation",
        "tags": [
          "evidence",
          "investigations"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}": {
      "patch": {
        "operationId": "update_investigation_api_v1_investigations__investigation_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Investigation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Investigation Api V1 Investigations  Investigation Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Investigation",
        "tags": [
          "evidence",
          "investigations"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/events": {
      "post": {
        "operationId": "add_investigation_event_api_v1_investigations__investigation_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Investigation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Investigation Event Api V1 Investigations  Investigation Id  Events Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Investigation Event",
        "tags": [
          "evidence",
          "investigations"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/feedback": {
      "post": {
        "operationId": "add_feedback_api_v1_investigations__investigation_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Investigation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-ID",
            "required": true,
            "schema": {
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Feedback Api V1 Investigations  Investigation Id  Feedback Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Feedback",
        "tags": [
          "evidence",
          "investigations"
        ]
      }
    },
...
  }
}